WIP basic workflow test for containers
This is a basic workflow test of the containers containing
the tasks that a typical user would run through to use them
and get familiar with ilab.

It is written in shell script because this basic workflow *is*
a shell workflow, run through step by step at a shell prompt
by a user.

Signed-off-by: Stef Walter <stefw@redhat.com>