e2e: Add minimal and help options
Add -m (minimal) and -h (help) options to the end-to-end test script.

I'm trying to get a config that's usable in a test env with no GPUs.
It doesn't need to produce a good result. I just want it to run all of
the steps in a minimal config. So far that's adding
`--num-instructions 1` for generate.

Signed-off-by: Russell Bryant <rbryant@redhat.com>